+++ Scripts/Items/Tradeable.cs
@@ -4,7 +4,7 @@ namespace Sands
 {
     public class Tradeable : Item
     {
-        public Tradeable(int ID, string itemName, double weight, double price) : base(ID, itemName, weight, price)
+        public Tradeable(int ID, string itemName, int weight, double price) : base(ID, itemName, weight, price)
         {
 
         }
